add config for vscode, vim; tune clang-format
