+++ .clang-format
@@ -6,11 +6,13 @@ AccessModifierOffset: -4
 AlignConsecutiveAssignments: false
 AlignConsecutiveDeclarations: false
 AlignOperands: true
-AlignTrailingComments: false
+AlignTrailingComments: true
 AlwaysBreakTemplateDeclarations: Yes
+AllowShortFunctionsOnASingleLine: Empty
+AllowShortBlocksOnASingleLine: Empty
 BraceWrapping:
   AfterObjCDeclaration: false
-  IndentBraces: false
+  IndentBraces: true
   AfterCaseLabel: false
   AfterClass: false
   AfterControlStatement: false
@@ -59,5 +61,5 @@ SpacesInConditionalStatement: false
 SpacesInCStyleCastParentheses: false
 SpacesInParentheses: false
 TabWidth: 4
-UseTab: Always
+UseTab: Never
 ...
