+++ compile_flags.txt
@@ -1,0 +1,6 @@
+-I
+godot-cpp/gdextension
+-I
+godot-cpp/gen/include
+-I
+godot-cpp/include
