+++ src/register_types.cpp
@@ -4,7 +4,7 @@
 #include <godot_cpp/core/defs.hpp>
 #include <godot_cpp/godot.hpp>
 
-#include "MaszynaTBrake.h"
+#include "MaszynaTBrake.hpp"
 
 using namespace godot;
